Rename external-api-usage-queries.ts to model-editor-queries.ts
