+++ extensions/ql-vscode/src/model-editor/model-editor-queries.ts
