+++ extensions/ql-vscode/test/vscode-tests/no-workspace/model-editor/external-api-usage-query.test.ts
@@ -1,7 +1,7 @@
 import {
   readQueryResults,
   runModelEditorQueries,
-} from "../../../../src/model-editor/external-api-usage-queries";
+} from "../../../../src/model-editor/model-editor-queries";
 import { createMockLogger } from "../../../__mocks__/loggerMock";
 import {
   DatabaseItem,
